requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
requests==2.32.5
psycopg2-binary==2.9.11